requirements.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
adjustText==1.1.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attr==0.3.2
attrs==23.2.0
braceexpand==0.1.7
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip==0.2.0
cmake==3.28.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.14.0
deepspeed==0.10.0
diffusers==0.20.0
dill==0.3.7
einops==0.6.1
fairscale==0.4.13
filelock==3.13.4
flash-attn==2.0.4
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.1
hjson==3.1.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==17.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.4
mmcv-full==1.7.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
MultiScaleDeformableAttention==1.0
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
palettable==3.3.3
pandas==2.2.0
peft==0.3.0
Pillow==10.0.0
platformdirs==4.2.0
protobuf==4.23.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.14
pyparsing==3.1.1
pyre-extensions==0.0.29
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytz==2024.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.1
sentencepiece==0.1.99
sentry-sdk==1.43.0
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2024.1.30
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1+cu117
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
webdataset==0.2.48
Werkzeug==3.0.1
xformers==0.0.20
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
